requirements.txt
pyserial
numpy
pandas
